Modularize firmware + resync full/headless trees
Split network/detect/scanner monoliths into focused modules and drop dead code. Reconcile drift so the trees differ only by control surface (web UI vs mesh): port BLE/identity/SD feature to headless, add drone bounds + heap guards, triangulation mutexes, hitCount overflow guard, HT-IE channel parse; remove WiFi-disable from battery saver. Both envs build; exhaustive cppcheck clean.

+++ Antihunter/headless/src/baseline.h
@@ -19,7 +19,7 @@ struct BaselineDevice {
     char name[32];
     bool isBLE;
     uint8_t channel;
-    uint16_t hitCount;
+    uint32_t hitCount;
     uint8_t checksum;
     bool dirtyFlag;
 } __attribute__((packed));
